refactor: narrow plugin host and split HTTP operations
