+++ packages/server/src/tools/admin/manage_operations.ts
@@ -30,7 +30,6 @@ import {
 } from "../../db/client";
 import type { Env } from "../../index";
 import { callTool as callProxyTool } from "../../mcp-proxy/client";
-import { resolveCredentialsByConnectionId } from "../../mcp-proxy/credential-resolver";
 import {
 	agentExistsInOrg,
 	resolveActingPrincipal,
@@ -51,6 +50,7 @@ import {
 	getOperationForConnection,
 	listOperations,
 } from "../../operations/connector-operations";
+import { executeHttpOperation } from "../../operations/execute-http-operation";
 import { validateOperationInput } from "../../operations/input-validation";
 import type {
 	AvailableOperation,
@@ -358,168 +358,6 @@ async function executeMcpToolInline(
 	} as Record<string, unknown>);
 }
 
-function buildResolvedUrl(
-	serverUrl: string,
-	pathTemplate: string,
-	input: Record<string, unknown>,
-): URL {
-	const pathValues =
-		input.path && typeof input.path === "object"
-			? (input.path as Record<string, unknown>)
-			: {};
-	const queryValues =
-		input.query && typeof input.query === "object"
-			? (input.query as Record<string, unknown>)
-			: {};
-	let path = pathTemplate;
-	for (const [key, value] of Object.entries(pathValues)) {
-		path = path.replaceAll(`{${key}}`, encodeURIComponent(String(value)));
-	}
-
-	const url = new URL(
-		path,
-		serverUrl.endsWith("/") ? serverUrl : `${serverUrl}/`,
-	);
-	for (const [key, value] of Object.entries(queryValues)) {
-		if (value === undefined || value === null) continue;
-		if (Array.isArray(value)) {
-			for (const item of value) url.searchParams.append(key, String(item));
-		} else {
-			url.searchParams.set(key, String(value));
-		}
-	}
-	return url;
-}
-
-function pickInterestingHeaders(
-	headers: Headers,
-): Record<string, unknown> | undefined {
-	const interestingHeaders = [
-		"content-type",
-		"x-ratelimit-remaining",
-		"x-ratelimit-reset",
-		"link",
-	];
-  const values = Object.fromEntries(
-    interestingHeaders
-      .map((header) => [header, headers.get(header)])
-			.filter(([, value]) => value !== null),
-  );
-  return Object.keys(values).length > 0 ? values : undefined;
-}
-
-async function executeHttpOperationInline(
-	runId: number,
-	organizationId: string,
-	connection: ConnectionRow,
-	operation: OperationDescriptor,
-	actionInput: Record<string, unknown>,
-): Promise<InlineExecutionResult> {
-	const sql = getDb();
-	if (operation.backend_config.backend !== "http_operation") {
-		return {
-			status: "failed",
-			error_message: "Invalid HTTP operation backend config",
-		};
-	}
-
-	const credentials = await resolveCredentialsByConnectionId(
-		connection.id,
-		organizationId,
-	);
-	if (!credentials) {
-		return {
-			status: "failed",
-			error_message: `No active OAuth credentials found for '${connection.connector_key}'.`,
-		};
-	}
-
-	const headersInput =
-		actionInput.headers && typeof actionInput.headers === "object"
-			? (actionInput.headers as Record<string, unknown>)
-			: {};
-	const headers = new Headers();
-	for (const [key, value] of Object.entries(headersInput)) {
-		if (
-			/^(authorization|host)$/i.test(key) ||
-			value === undefined ||
-			value === null
-		)
-			continue;
-		headers.set(key, String(value));
-	}
-	headers.set(
-		"Authorization",
-		`${credentials.tokenType} ${credentials.accessToken}`,
-	);
-
-	const body = actionInput.body;
-	let requestBody: string | undefined;
-	if (body !== undefined) {
-		requestBody = typeof body === "string" ? body : JSON.stringify(body);
-		if (
-			!headers.has("content-type") &&
-			typeof body === "object" &&
-			body !== null
-		) {
-			headers.set("content-type", "application/json");
-		}
-	}
-
-	const url = buildResolvedUrl(
-		operation.backend_config.serverUrl,
-		operation.backend_config.pathTemplate,
-		actionInput,
-	);
-
-	try {
-		const response = await fetch(url, {
-			method: operation.backend_config.method,
-			headers,
-			body: ["GET", "HEAD"].includes(operation.backend_config.method)
-				? undefined
-				: requestBody,
-			redirect: "manual",
-		});
-
-		const text = await response.text();
-		let parsedBody: unknown = text;
-		try {
-			parsedBody = text ? JSON.parse(text) : null;
-		} catch {
-			// Keep as text
-		}
-		const output = { body: parsedBody } as Record<string, unknown>;
-		const metadata: Record<string, unknown> = {
-			http_status: response.status,
-		};
-		const headerMetadata = pickInterestingHeaders(response.headers);
-		if (headerMetadata) {
-			metadata.response_headers = headerMetadata;
-			const rateLimits = Object.fromEntries(
-				Object.entries(headerMetadata).filter(([key]) =>
-					key.startsWith("x-ratelimit"),
-				),
-			);
-			if (Object.keys(rateLimits).length > 0) metadata.rate_limits = rateLimits;
-			if (headerMetadata.link)
-				metadata.pagination = { link: headerMetadata.link };
-		}
-
-		if (!response.ok) {
-			const errorText =
-				typeof parsedBody === "string" ? parsedBody : `HTTP ${response.status}`;
-			await sql`UPDATE runs SET status = 'failed', completed_at = NOW(), action_output = ${sql.json(output)}, error_message = ${errorText} WHERE id = ${runId} AND organization_id = ${organizationId}`;
-			return { status: "failed", error_message: errorText };
-		}
-
-		await sql`UPDATE runs SET status = 'completed', completed_at = NOW(), action_output = ${sql.json(output)} WHERE id = ${runId} AND organization_id = ${organizationId}`;
-		return { status: "completed", output, metadata };
-	} catch (error) {
-		return failRunInline(runId, organizationId, getErrorMessage(error));
-	}
-}
-
 async function executeOperationInline(
 	runId: number,
 	organizationId: string,
@@ -549,7 +387,7 @@ async function executeOperationInline(
 			actionInput,
 		);
   }
-	return executeHttpOperationInline(
+	return executeHttpOperation(
 		runId,
 		organizationId,
 		connection,
